+++ server/lsys-app-sender/src/dao/sender_mailer/record.rs
@@ -144,7 +144,7 @@ impl MailRecord {
             qb.push(") as t");
         }
 
-        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await? as u64;
+        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await?;
 
         Ok(query.finalize(count))
     }
