+++ server/lsys-app-sender/src/dao/sender_smser/record.rs
@@ -147,7 +147,7 @@ impl SmsRecord {
             qb.push(") as t");
         }
 
-        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await? as u64;
+        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await?;
 
         Ok(query.finalize(count))
     }
