+++ server/lsys-app/src/dao/app_notify/record.rs
@@ -487,7 +487,7 @@ impl AppNotifyRecord {
             query.push_limit(&mut qb);
             qb.push(") as t");
         }
-        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await? as u64;
+        let count = qb.build_query_scalar::<i64>().fetch_one(&self.db).await?;
         Ok(query.finalize(count))
     }
     //消息列表
