+++ server/lsys-logger/src/dao/change_log.rs
@@ -192,7 +192,7 @@ impl ChangeLoggerDao {
             .build_query_scalar()
             .fetch_one(&self.db)
             .await
-            .unwrap_or(0i64) as u64;
+            .unwrap_or(0i64);
         Ok(query.finalize(count))
     }
 }
